fix(v1/user): getGamesId関数をUserクラスに内包化
